+++ glibc/math/__Float32x4_t.c
@@ -1,0 +1,8 @@
+#include <math.h>
+
+int main(void)
+{
+	__Float32x4_t f32x4;
+	(void)f32x4;
+	return 0;
+}