+++ glibc/math/__Float64x2_t.c
@@ -1,0 +1,8 @@
+#include <math.h>
+
+int main(void)
+{
+	__Float64x2_t f64x2;
+	(void)f64x2;
+	return 0;
+}